refactor vtx handling
