+++ src/io/msp.c
@@ -27,7 +27,7 @@ enum {
   MSP_REBOOT_COUNT,
 };
 
-extern uint8_t msp_vtx_detected;
+extern bool msp_vtx_detected;
 extern vtx_settings_t vtx_actual;
 extern char msp_vtx_band_letters[VTX_BAND_MAX];
 extern uint8_t msp_vtx_band_is_factory[VTX_BAND_MAX];
@@ -63,6 +63,16 @@ static void msp_write_uint32_t(uint8_t *data, uint32_t val) {
   data[1] = val >> 8;
   data[2] = val >> 16;
   data[3] = val >> 24;
+}
+
+static void msp_check_vtx_detected(msp_t *msp) {
+  if (msp_vtx_detected || msp->device != MSP_DEVICE_VTX)
+    return;
+
+  if (vtx_actual.power_table.levels == 0)
+    return;
+
+  msp_vtx_detected = true;
 }
 
 static void msp_process_serial_cmd(msp_t *msp, msp_magic_t magic, uint16_t cmd, uint8_t *payload, uint16_t size) {
@@ -333,6 +343,7 @@ static void msp_process_serial_cmd(msp_t *msp, msp_magic_t magic, uint16_t cmd, 
 #endif
 #ifdef USE_VTX
   case MSP_VTX_CONFIG: {
+    msp_check_vtx_detected(msp);
     msp_vtx_send_config_reply(msp, magic);
     break;
   }
@@ -404,6 +415,7 @@ static void msp_process_serial_cmd(msp_t *msp, msp_magic_t magic, uint16_t cmd, 
       vtx_set(settings);
     }
 
+    msp_check_vtx_detected(msp);
     msp_send_reply(msp, magic, cmd, NULL, 0);
     break;
   }
@@ -505,7 +517,7 @@ static void msp_process_serial_cmd(msp_t *msp, msp_magic_t magic, uint16_t cmd, 
   case MSP_EEPROM_WRITE: {
 #ifdef USE_VTX
     if (msp->device == MSP_DEVICE_VTX) {
-      msp_vtx_detected = 1;
+      msp_check_vtx_detected(msp);
     } else
 #endif
         if (!flags.arm_state && msp->device != MSP_DEVICE_SPI_RX) {
