+++ src/io/vtx.c
@@ -9,9 +9,6 @@
 #include "driver/gpio.h"
 #include "driver/serial.h"
 #include "driver/time.h"
-#include "driver/vtx/msp.h"
-#include "driver/vtx/sa.h"
-#include "driver/vtx/tramp.h"
 #include "flight/control.h"
 #include "rx/rx.h"
 #include "rx/unified_serial.h"
@@ -33,28 +30,10 @@ static uint8_t apply_tries = 0;
 static uint32_t vtx_delay_start = 0;
 static uint32_t vtx_delay_ms = 100;
 
-extern uint8_t smart_audio_detected;
-extern smart_audio_settings_t smart_audio_settings;
-
-vtx_detect_status_t vtx_smart_audio_update(vtx_settings_t *actual);
-void smart_audio_set_frequency(vtx_band_t band, vtx_channel_t channel);
-void smart_audio_set_power_level(vtx_power_level_t power);
-void smart_audio_set_pit_mode(vtx_pit_mode_t pit_mode);
-
-extern uint8_t tramp_detected;
-extern tramp_settings_t tramp_settings;
-
-vtx_detect_status_t vtx_tramp_update(vtx_settings_t *actual);
-void tramp_set_frequency(vtx_band_t band, vtx_channel_t channel);
-void tramp_set_power_level(vtx_power_level_t power);
-void tramp_set_pit_mode(vtx_pit_mode_t pit_mode);
-
-extern uint8_t msp_vtx_detected;
-
-vtx_detect_status_t vtx_msp_update(vtx_settings_t *actual);
-void msp_vtx_set_frequency(vtx_band_t band, vtx_channel_t channel);
-void msp_vtx_set_power_level(vtx_power_level_t power);
-void msp_vtx_set_pit_mode(vtx_pit_mode_t pit_mode);
+static const vtx_device_t *vtx_device = NULL;
+extern const vtx_device_t msp_vtx_device;
+extern const vtx_device_t smart_audio_vtx_device;
+extern const vtx_device_t tramp_vtx_device;
 
 const uint16_t frequency_table[VTX_BAND_MAX][VTX_CHANNEL_MAX] = {
     {5865, 5845, 5825, 5805, 5785, 5765, 5745, 5725}, // VTX_BAND_A
@@ -89,25 +68,6 @@ vtx_power_level_t vtx_power_level_index(vtx_power_table_t *power_table, uint16_t
   return VTX_POWER_LEVEL_1;
 }
 
-static vtx_detect_status_t vtx_update_protocol(vtx_protocol_t proto, vtx_settings_t *actual) {
-  switch (proto) {
-  case VTX_PROTOCOL_TRAMP:
-    return vtx_tramp_update(actual);
-
-  case VTX_PROTOCOL_SMART_AUDIO:
-    return vtx_smart_audio_update(actual);
-
-  case VTX_PROTOCOL_MSP_VTX:
-    return vtx_msp_update(actual);
-
-  case VTX_PROTOCOL_INVALID:
-  case VTX_PROTOCOL_MAX:
-    return VTX_DETECT_ERROR;
-  }
-
-  return VTX_DETECT_ERROR;
-}
-
 void vtx_init() {
   vtx_settings.detected = VTX_PROTOCOL_INVALID;
   vtx_delay_start = time_millis();
@@ -140,7 +100,7 @@ static void vtx_update_fpv_pin() {
 }
 
 static bool vtx_detect_protocol() {
-  static vtx_protocol_t protocol_to_check = VTX_PROTOCOL_MSP_VTX;
+  static vtx_protocol_t protocol_to_check = VTX_PROTOCOL_SMART_AUDIO;
   static uint8_t protocol_is_init = 0;
 
   if (vtx_settings.detected) {
@@ -164,27 +124,26 @@ static bool vtx_detect_protocol() {
   if (!protocol_is_init) {
     switch (protocol_to_check) {
     case VTX_PROTOCOL_TRAMP:
-      serial_tramp_init();
+      vtx_device = &tramp_vtx_device;
       break;
 
     case VTX_PROTOCOL_SMART_AUDIO:
-      serial_smart_audio_init();
+      vtx_device = &smart_audio_vtx_device;
       break;
 
     case VTX_PROTOCOL_MSP_VTX:
-      serial_msp_vtx_init();
+      vtx_device = &msp_vtx_device;
       break;
 
-    case VTX_PROTOCOL_INVALID:
-    case VTX_PROTOCOL_MAX:
-      break;
+    default:
+      vtx_device = NULL;
+      return false;
     }
+    vtx_device->init();
     protocol_is_init = 1;
-    return false;
   }
 
-  const vtx_detect_status_t status = vtx_update_protocol(protocol_to_check, &vtx_actual);
-
+  const vtx_detect_status_t status = vtx_device->update(&vtx_actual);
   if (status == VTX_DETECT_SUCCESS) {
     // detect success, save detected proto
     vtx_settings.protocol = protocol_to_check;
@@ -202,7 +161,6 @@ static bool vtx_detect_protocol() {
       }
     }
   }
-
   return false;
 }
 
@@ -220,26 +178,10 @@ static bool vtx_update_frequency() {
     return true;
   }
 
-  switch (vtx_settings.detected) {
-  case VTX_PROTOCOL_TRAMP:
-    tramp_set_frequency(vtx_settings.band, vtx_settings.channel);
-    break;
-
-  case VTX_PROTOCOL_SMART_AUDIO:
-    smart_audio_set_frequency(vtx_settings.band, vtx_settings.channel);
-    break;
-
-  case VTX_PROTOCOL_MSP_VTX:
-    msp_vtx_set_frequency(vtx_settings.band, vtx_settings.channel);
-    break;
-
-  case VTX_PROTOCOL_INVALID:
-  case VTX_PROTOCOL_MAX:
-    break;
+  if (vtx_device->set_frequency(vtx_settings.band, vtx_settings.channel)) {
+    vtx_delay_ms = 10;
+    apply_tries++;
   }
-
-  apply_tries++;
-  vtx_delay_ms = 10;
   return false;
 }
 
@@ -256,27 +198,10 @@ static bool vtx_update_powerlevel() {
     return true;
   }
 
-  switch (vtx_settings.detected) {
-  case VTX_PROTOCOL_TRAMP:
-    tramp_set_power_level(vtx_settings.power_level);
-    break;
-
-  case VTX_PROTOCOL_SMART_AUDIO:
-    smart_audio_set_power_level(vtx_settings.power_level);
-    break;
-
-  case VTX_PROTOCOL_MSP_VTX:
-    msp_vtx_set_power_level(vtx_settings.power_level);
-    break;
-
-  case VTX_PROTOCOL_INVALID:
-  case VTX_PROTOCOL_MAX:
-    break;
+  if (vtx_device->set_power_level(vtx_settings.power_level)) {
+    vtx_delay_ms = 10;
+    apply_tries++;
   }
-
-  apply_tries++;
-  vtx_delay_ms = 10;
-
   return false;
 }
 
@@ -293,27 +218,10 @@ static bool vtx_update_pitmode() {
     return true;
   }
 
-  switch (vtx_settings.detected) {
-  case VTX_PROTOCOL_TRAMP:
-    tramp_set_pit_mode(vtx_settings.pit_mode);
-    break;
-
-  case VTX_PROTOCOL_SMART_AUDIO:
-    smart_audio_set_pit_mode(vtx_settings.pit_mode);
-    break;
-
-  case VTX_PROTOCOL_MSP_VTX:
-    msp_vtx_set_pit_mode(vtx_settings.pit_mode);
-    break;
-
-  case VTX_PROTOCOL_INVALID:
-  case VTX_PROTOCOL_MAX:
-    break;
+  if (vtx_device->set_pit_mode(vtx_settings.pit_mode)) {
+    vtx_delay_ms = 10;
+    apply_tries++;
   }
-
-  apply_tries++;
-  vtx_delay_ms = 10;
-
   return false;
 }
 
@@ -348,8 +256,11 @@ void vtx_update() {
   if (!vtx_detect_protocol()) {
     return;
   }
+  if (vtx_device == NULL) {
+    return;
+  }
 
-  const vtx_detect_status_t status = vtx_update_protocol(vtx_settings.detected, &vtx_actual);
+  const vtx_detect_status_t status = vtx_device->update(&vtx_actual);
   if (status < VTX_DETECT_SUCCESS) {
     // we are in wait or error state, do nothing
     return;
@@ -385,12 +296,6 @@ void vtx_set(vtx_settings_t *vtx) {
     vtx_settings.protocol = vtx->protocol;
     vtx_settings.magic = 0xFFFF;
     vtx_settings.power_table.levels = 0;
-
-    smart_audio_settings.version = 0;
-    smart_audio_detected = 0;
-    tramp_settings.freq_min = 0;
-    tramp_detected = 0;
-    msp_vtx_detected = 0;
   } else {
     vtx_settings.magic = VTX_SETTINGS_MAGIC;
     memcpy(&vtx_settings.power_table, &vtx->power_table, sizeof(vtx_power_table_t));
